Test: example of a top-level project system test script
The idea here is that it can find and run all the unit tests for all
packages in the project and then run some arbitrary functional or
system tests that do things like run binaries and see what they do.

We add a second binary to show how that works. Each `.go` file that
defines a `package main ... func main() { ... }` needs to be in its
own directory. (The filename within the directory doesn't matter.)
See the discussion of package `main` in [1].

[1]: https://golang.org/cmd/go/#hdr-Description_of_package_lists